refactor: restructure project to src-based architecture
- move core folders into src/
- group logic by modules (auth, location, etc.)
- separate configs, hooks, utils
- no business logic changes
