+++ app/(app)/index.tsx
@@ -1,6 +1,6 @@
-import BottomWayMenu from "@/components/_views/BottomWayMenu";
-import Map from "@/components/_views/Map";
-import { auth } from "@/configs/firebaseConfig";
+import BottomWayMenu from "@/src/components/_views/BottomWayMenu";
+import Map from "@/src/components/_views/Map";
+import { auth } from "@/src/configs/firebaseConfig";
 import { signOut } from "firebase/auth";
 import React from "react";
 import { Alert, Text, View } from "react-native";
@@ -19,13 +19,19 @@ export default function Index() {
   return (
     <View style={{ flex: 1, justifyContent: "center" }}>
       <Button
-        style={{position:"absolute", zIndex:4, backgroundColor: "gray", padding: 10, borderRadius: 5 }}
+        style={{
+          position: "absolute",
+          zIndex: 4,
+          backgroundColor: "gray",
+          padding: 10,
+          borderRadius: 5,
+        }}
         onPress={handlePress}
       >
         <Text style={{ color: "white", fontSize: 16 }}>Out</Text>
       </Button>
-      <Map/>
-      
+      <Map />
+
       <BottomWayMenu />
     </View>
   );
