+++ src/auth/useContinueWithGoogle.ts
@@ -1,5 +1,5 @@
-import { auth } from "@/configs/firebaseConfig";
-import { LogService } from "@/utils/LogService";
+import { auth } from "@/src/configs/firebaseConfig";
+import { LogService } from "@/src/utils/LogService";
 import { GoogleSignin } from "@react-native-google-signin/google-signin";
 import { GoogleAuthProvider, signInWithCredential } from "firebase/auth";
