+++ src/components/_models/MapModel.ts
