+++ src/components/_views/CarsSwiper.tsx
@@ -1,4 +1,4 @@
-import { Images } from "@/constants/images";
+import { Images } from "@/src/constants/images";
 import React from "react";
 import { Image, Text, View } from "react-native";
 import normalize from "react-native-normalize";
