+++ src/components/_views/RegistrationAnimatedView.tsx
