+++ src/components/_viewsModels/BottomWayMenuViewModel.ts
@@ -1,4 +1,4 @@
-import { RootState } from "@/store/store";
+import { RootState } from "@/src/store/store";
 import { useState } from "react";
 import { useSelector } from "react-redux";
