+++ src/components/_viewsModels/RegistrationAnimatedViewModel.ts
