+++ src/store/Slices/map/mapSlice.ts
@@ -1,4 +1,4 @@
-import { RouteInfo } from "@/types/routeTypes";
+import { RouteInfo } from "@/src/types/routeTypes";
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { initialState } from "./mapInterfase";
